SCRUM-71: completed and code corrections

+++ src/main/java/com/utn/space/venueaapi/Repository/ConsumersRepository.java
@@ -5,6 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-
-public interface ConsumersRepository extends JpaRepository<Consumers,Integer> {
+public interface ConsumersRepository extends JpaRepository<Consumers,Long> {
 }
