+++ src/main/java/com/utn/space/venueaapi/Repository/SpaceImagesRepository.java
@@ -4,6 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-
-public interface SpaceImagesRepository extends JpaRepository<SpaceImagesRepository,Integer> {
+public interface SpaceImagesRepository extends JpaRepository<SpaceImagesRepository,Long> {
 }
