+++ src/main/java/com/utn/space/venueaapi/Repository/SpacesRepository.java
@@ -5,6 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-
-public interface SpacesRepository extends JpaRepository<Spaces,Integer> {
+public interface SpacesRepository extends JpaRepository<Spaces,Long> {
 }
